On branch raspberrypi Changes to be committed:   (use "git reset HEAD <file>..." to unstage)
modified:   install_workspace.sh
new file:   usr/bin/backup_evry
modified:   usr/bin/backup_local
new file:   usr/bin/backup_loria
modified:   usr/bin/backup_remote
new file:   usr/bin/batterylevel
new file:   usr/bin/l33t
new file:   usr/bin/tmpfs
modified:   usr/bin/upd
modified:   usr/share/tex/macro.tex
new file:   usr/share/tex/macro_these.tex
modified:   usr/var/backup/excluded
new file:   usr/var/backup/excluded_duplicity
new file:   usr/var/backup/rsync/excluded_loria
modified:   usr/var/bibtex/references.bib
new file:   usr/var/color/color

+++ install_workspace.sh
@@ -1,5 +1,7 @@
 #!/bin/bash
 
+source usr/var/color/color
+
 locdir=$(pwd)
 listelements=$locdir/* #ne prends que les fichiers visibles
 
@@ -15,5 +17,18 @@ do
 	link=${file/$suffixe/"$HOME/."}
 	echo $file
 	echo "   -> $link"
-	ln -s $file $link
+  if [ -r $link ]
+  then
+    if [ "$1" == "--force" ]
+    then
+      echo -e "${orange}   -> écrase ancienne version${neutre}"
+      rm $link
+      ln -s $file $link
+    else
+      echo -e "${rougefonce}   -> déjà présent !${neutre}"
+    fi
+  else
+    echo -e "${vertfonce}   -> ok !${neutre}"
+    ln -s $file $link
+  fi
 done
